Split up `agent.py`
It's getting unwieldy to work with. Move all code that is specific to a
particular type of function into dedicated files.
